requirements.txt
streamlit
qrcode
pillow
numpy